+++ frontend/components/HTMLContent/index.jsx
@@ -1,35 +1,36 @@
 import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
-import { css } from 'glamor';
 import { HtmlSanitizer } from '@shopgate/engage/components';
-import { useNavigation } from '@shopgate/engage/core';
-import { themeConfig } from '@shopgate/pwa-common/helpers/config';
+import { useNavigation } from '@shopgate/engage/core/hooks';
+import { makeStyles } from '@shopgate/engage/styles';
 
-const contentStyles = css({
-  lineHeight: 1.7,
-  overflow: 'hidden',
-  wordBreak: ['break-all', 'break-word'],
-  hyphens: 'auto',
-  ' ul': {
-    listStyle: 'disc',
+const useStyles = makeStyles()(theme => ({
+  content: {
+    lineHeight: 1.7,
+    overflow: 'hidden',
+    wordBreak: ['break-all', 'break-word'],
+    hyphens: 'auto',
+    '& ul': {
+      listStyle: 'disc',
+    },
+    '& ol': {
+      listStyle: 'decimal',
+    },
+    '& ul, & ol': {
+      margin: '.75em 0',
+      paddingLeft: '1.2em',
+    },
+    '& a': {
+      color: theme.palette.primary.main,
+      margin: '-.35em',
+      padding: '.35em',
+      position: 'relative',
+    },
+    '& p': {
+      marginTop: 0,
+    },
   },
-  ' ol': {
-    listStyle: 'decimal',
-  },
-  ' ul, ol': {
-    margin: '.75em 0',
-    paddingLeft: '1.2em',
-  },
-  ' a': {
-    color: themeConfig.colors.primary,
-    margin: '-.35em',
-    padding: '.35em',
-    position: 'relative',
-  },
-  ' p': {
-    marginTop: 0,
-  },
-});
+}));
 
 /**
  * HTMLContent
@@ -37,6 +38,7 @@ const contentStyles = css({
  * @returns {JSX}
  */
 const HTMLContent = ({ children, contentId, processStyles }) => {
+  const { classes } = useStyles();
   const { push } = useNavigation();
 
   const handleClick = useCallback((pathname, target) => {
@@ -47,7 +49,7 @@ const HTMLContent = ({ children, contentId, processStyles }) => {
   }, [push]);
 
   return (
-    <div className={contentStyles}>
+    <div className={classes.content}>
       <HtmlSanitizer
         processStyles={processStyles}
         settings={{